--- test/ui-e2e/.gitignore
@@ -1,1 +1,0 @@
-html
