--- test/ui-e2e/fixtures/sample.test.ts
@@ -1,7 +1,0 @@
-import { expect, it } from 'vitest'
-
-it('add', () => {
-  // eslint-disable-next-line no-console
-  console.log('log test')
-  expect(1 + 1).toEqual(2)
-})
