--- test/ui-e2e/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@vitest/test-ui-e2e",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "test": "playwright test",
-    "test-fixtures": "vitest"
-  },
-  "devDependencies": {
-    "@playwright/test": "^1.39.0",
-    "execa": "^6.1.0",
-    "vitest": "workspace:*"
-  }
-}
