--- test/ui-e2e/vitest.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    dir: './fixtures',
-  },
-})
