+++ test/ui/.gitignore
@@ -1,1 +1,1 @@
-fixtures/html
+html
