--- test/ui/fixtures/vitest.config.ts
@@ -1,4 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-})
