+++ test/ui/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "@vitest/test-ui",
+  "name": "@vitest/test-ui-e2e",
   "type": "module",
   "private": true,
   "scripts": {
-    "test": "vitest run"
+    "test": "playwright test",
+    "test-fixtures": "vitest"
   },
   "devDependencies": {
+    "@playwright/test": "^1.39.0",
     "execa": "^6.1.0",
-    "playwright-chromium": "^1.39.0",
     "vitest": "workspace:*"
   }
 }
