--- test/ui/shim.d.ts
@@ -1,4 +1,0 @@
-declare module 'kill-port' {
-  const kill: (port: number) => Promise<void>
-  export default kill
-}
